Make the failed message bytes available in a CRCError exception.
When a serial CRC validation fails, the exception is constructed with
a descriptive text as its only argument.  To aid in debugging, pass in
the original failing message instead.  Make sure the string
representation of the exception stays the same, by adding a __str__()
method like in the library's other exceptions.

Via https://github.com/AdvancedClimateSystems/uModbus/pull/113